Add remove this device option in Accounts Page. Move AccountsPage components to separated folder
